Stream edit page tweaks (#984)
* Rename StreamPage/Show to StreamPage/Edit

* Drop redundant condition

We already check if stream&.id is set in the upper part of
the component.

* Move PreviewView’s description to the stream edit page

We don’t want it to be part of the actual component. This way we can
reuse PreviewView “as is” in the StreamPage/View component.

* Refactor PreviewView real quick to use styled components

* Fix a margin spill

Don’t understand code? Remove it. No strings attached. XD

* Make TOCPage’s title optional

* Drop `disabled` prop optional

C’moan guyz!

* Make setting edited stream to null reset it properly

* Keep location’s `search` while scrolling

* Add `updating` to userpage stream module

We know when we’re `fetching` and now we also know when we’re `fetching`
cause of an update.