Better navigation blocker
`history.block` can only be called once at a time. This makes the previous `usePreventNavigatingAway` limited to a single block per route. We have 2, in most cases (stream page and the one in `Globals`).

I set up a blocker registry and make `history.block` go through it and use the first eligible blocker. If it doesn't find any, it idles.

Too tired to write more. Cheers!