Fall back onto the Not Found page
After the recent change to the `app/index` we've made the `StreamPage` and the `ProjectPage` responsible for pretty much all traffic going towards `/hub/streams/*` and `/hub/projects/:id/*` respectively.

In this commit I make sure invalid links render the 404 page.

Note that this does not solve the remaining issues with links pointing to non-existing projects. Work on those is pending.