Fix tests
First of all, we tell `ts-jest` to transform js files, too. Secondly, we tell the transformer include `query-string` module as part of the process.

Simple as that.

Ignoring `<rootDir>/node_modules/…` is a bit different than ignoring `/node_modules/…`. It should be safe though. I assume that the only modules that have their own `node_modules` folders are es6 modules and we'd have to transform them anyway.