Extract drafting logic from the project editor
This will make it possible to reuse the "init → abandon → teardown" lifecycle to other draftable things in the future.

In the next iteration I'll apply it to the stream editor, move it out from shared/ and refactor. There are things in the stream editor that don't belong there, too. It needs love!