Combine new and edit stream pages
Changing from "new" being a separate component to just using the edit page for both editing and creating.

Important: the aim here is to avoid unnecessary unmounts of the `useInitDraft` hook which'd then prevent redundant resource loading immediately after creation.

As a side effect of this change there's a new hook – `useInitDraft2` – which, unlike `useInitDraft`, takes only the entity id as an argument. It simplifies things. We then explicitly use `assign` to assign an instance of the entity when it's ready.

It's good.